+++ changelog.md
@@ -6,6 +6,7 @@
 - Root elo rankings pulled from Rootelo
 - Elo Ranking displayed in the stats section of the profile page
 - Duplicate Factions in the Forge so you can make changes without losing an old design
+- new /card command for the discord bot to lookup cards
 
 ### Improvements
 - Auto de-select winner if a score less than 30 is recorded